requirements.txt
Flask
requests
openai
google-generativeai
python-dotenv
langchain
huggingface_hub
langchain_google_genai